IGAPP-646: move Platform to PlatformContext & update snapshot
