+++ web-ts/src/services/wrapDisplayName.ts
@@ -2,7 +2,7 @@ import { ComponentType } from 'react'
 
 const getDisplayName = Component => Component.displayName || Component.name || typeof Component
 
-const wrapDisplayName = <Props>(Component: ComponentType<Props>, hocName: string) => {
+const wrapDisplayName = <Props>(Component: ComponentType<Props>, hocName: string): string => {
   return `${hocName}(${getDisplayName(Component)})`
 }
